requirements.txt
aiogram~=3.3.0
SQLAlchemy~=2.0.25
python-dotenv~=1.0.1
python-json-logger~=2.0.7
asyncpg~=0.29.0